+++ cluster_tools/test.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+set -eEuo pipefail
+
+echo "test.sh is not available for cluster_tools yet"