+++ webknossos/tests/test_annotation.py
@@ -8,6 +8,8 @@ from webknossos.geometry import BoundingBox, Vec3Int
 
 from .constants import TESTDATA_DIR, TESTOUTPUT_DIR
 
+pytestmark = [pytest.mark.with_vcr]
+
 
 def test_annotation_from_zip_file() -> None: