+++ wkcuber/Changelog.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/) `MAJOR.MIN
 For upgrade instructions, please check the respective *Breaking Changes* sections.
 
 ## Unreleased
-[Commits](https://github.com/scalableminds/webknossos-libs/compare/v0.9.14...HEAD)
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v0.9.18...HEAD)
 
 ### Breaking Changes
 
@@ -16,6 +16,22 @@ For upgrade instructions, please check the respective *Breaking Changes* section
 ### Changed
 
 ### Fixed
+
+
+## [0.9.18](https://github.com/scalableminds/webknossos-libs/releases/tag/v0.9.18) - 2022-04-06
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v0.9.17...v0.9.18)
+
+
+## [0.9.17](https://github.com/scalableminds/webknossos-libs/releases/tag/v0.9.17) - 2022-04-05
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v0.9.16...v0.9.17)
+
+
+## [0.9.16](https://github.com/scalableminds/webknossos-libs/releases/tag/v0.9.16) - 2022-03-31
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v0.9.15...v0.9.16)
+
+
+## [0.9.15](https://github.com/scalableminds/webknossos-libs/releases/tag/v0.9.15) - 2022-03-31
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v0.9.14...v0.9.15)
 
 
 ## [0.9.14](https://github.com/scalableminds/webknossos-libs/releases/tag/v0.9.14) - 2022-03-24